Refactored initial tooltip logic from 'TerrainControlDock' to 'CustomContentLoader' for better tooltip handling
